refactor: move sidebar to separate widget.
