+++ client/src/widgets/chat/ChatPageLayout.vue
