+++ client/src/widgets/sidebar/model/sidebarStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia";
 import { ref } from "vue";
 
-export const useChatStore = defineStore("chatStore", () => {
+export const sidebarStore = defineStore("sidebarStore", () => {
     const isProfileContextOpen = ref(false);
 
     return { isProfileContextOpen };
